requirements.txt
streamlit
faiss-cpu
sentence-transformers
pandas
numpy
ollama
